feat: Adding authorization by role logic

+++ services/service-3/src/controllers/fieldController.js
@@ -13,7 +13,6 @@ import {
 // DELETE: Field
 // Tambahkan validasi jangan lupa dan catch errror
 
-
 export const getAllFields = async (req, res) => {
   try {
     const fields = await findAllFields();
@@ -43,17 +42,27 @@ export const getFieldById = async (req, res) => {
 
 export const createFieldController = async (req, res) => {
   try {
-    const { owner_id, name, type, address, city, status } = req.body;
-    console.log(owner_id, name, type, address, city, status);
+    const role = req.headers["x-user-role"];
+    console.log(role);
 
-    if (!owner_id || !name || !type || !address || !city) {
+    if (role === "user") {
+      return res.status(403).json({ message: "Forbidden. Only owners can create field" });
+    }
+
+    const ownerId = req.headers["x-user-id"];
+    console.log("owner id:", ownerId);
+
+    const { name, type, address, city, status } = req.body;
+    console.log(name, type, address, city, status);
+
+    if (!name || !type || !address || !city) {
       return res
         .status(400)
         .json({ message: "Field wajib diisi:  name, type, address, city," });
     }
 
     const newField = await createField({
-      owner_id,
+      owner_id: ownerId,
       name,
       type,
       address,
@@ -75,7 +84,25 @@ export const createFieldController = async (req, res) => {
 
 export const updateFieldController = async (req, res) => {
   try {
+    const role = req.headers["x-user-role"];
+    console.log(role);
+
+    if (role === "user") {
+      return res.status(403).json({ message: "Forbidden. Only owners can create field" });
+    }
+
+    const ownerId = req.headers["x-user-id"];
+    console.log("owner id:", ownerId);
+
     const { id } = req.params;
+
+    const field = await findFieldById(id);
+
+    if (ownerId !== field.owner_id) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden. You're not the owner" });
+    }
 
     const updatedField = await updateField(id, req.body);
 
@@ -93,7 +120,25 @@ export const updateFieldController = async (req, res) => {
 
 export const deleteFieldController = async (req, res) => {
   try {
+    const role = req.headers["x-user-role"];
+    console.log(role);
+
+    if (role === "user") {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
+    const ownerId = req.headers["x-user-id"];
+    console.log("owner id:", ownerId);
+
     const { id } = req.params;
+
+    const field = await findFieldById(id);
+
+    if (ownerId !== field.owner_id) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden you're not the owner" });
+    }
 
     const deletedField = await deleteField(id);
